fiw IU

+++ app/src/main/java/com/example/myfamilyhandbook/NoteManager.kt
@@ -1,18 +1,18 @@
 package com.example.myfamilyhandbook
 
-import android.text.Editable
+
 
 object NotesManager {
     private val notes = mutableListOf<Note>()
 
-    fun addNote(name: String, age: Int) {
+     fun addNote(name: String, age: Int) {
         val id = if (notes.isEmpty()) {
            1
         } else {
             notes.last().id + 1
         }
 
-        notes.add(Note(id as Int, name, age))
+        notes.add(Note(id , name, age))
     }
 
     fun getNotes(): List<Note> {
